[TensorExpr] Nuke `Function` class and directly use `Tensor` instead.
`Tensor` has been a view into a `Function` that was supposed to be used
for a more general case when we have multiple computations over the same
domain (aka multiple output functions). We have never got to a point
where we need this and now have other ideas in mind on how to support
this case if need be. For now, let's just nuke `Function` to reduce the
overall system complexity.

The change should not affect any existing behavior.

ghstack-source-id: 873d452fda95def24486e70c5a20e07676901bc6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45936